Use layer_slug instead of layer_name
